hide some private structures
